# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notify Module",
        "contact": {
            "email": "pvalente@onesource.pt"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/type/{type_key}": {
            "get": {
                "tags": [
                    "Type"
                ],
                "summary": "Show a Type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type_key",
                        "in": "path",
                        "description": "Type Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "LANG-CODE",
                        "in": "header",
                        "description": "User Language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "LANG-CODE-DEFAULT",
                        "in": "header",
                        "description": "Entity default Language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show the Type data",
                        "schema": {
                            "$ref": "#/definitions/typeReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Type not Found",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve Type",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Type"
                ],
                "summary": "Update a Type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Body",
                        "in": "body",
                        "description": "Type Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/typeCreate"
                        }
                    },
                    {
                        "name": "type_key",
                        "in": "path",
                        "description": "Type Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated Type",
                        "schema": {
                            "$ref": "#/definitions/typeReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Type not Found",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to update Type",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Type"
                ],
                "summary": "Delete an Type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type_key",
                        "in": "path",
                        "description": "Type Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/typeDeleteReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Type not Found",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to delete Type",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    }
                }
            }
        },
        "/type": {
            "post": {
                "tags": [
                    "Type"
                ],
                "summary": "Create a Type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Body",
                        "in": "body",
                        "description": "Type Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/typeCreate"
                        }
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the newly created Type",
                        "schema": {
                            "$ref": "#/definitions/typeReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Type not found",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to store Type",
                        "schema": {
                            "$ref": "#/definitions/typeErrorDefault"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "typeErrorDefault": {
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "typeTranslations": {
            "type": "object",
            "allOf": [
                {
                    "required": [
                        "language_code",
                        "subject",
                        "header",
                        "content",
                        "footer"
                    ],
                    "properties": {
                        "language_code": {
                            "type": "string",
                            "format": "string"
                        },
                        "name": {
                            "type": "string",
                            "format": "string"
                        },
                        "subject": {
                            "type": "string",
                            "format": "string"
                        },
                        "header": {
                            "type": "string",
                            "format": "string"
                        },
                        "content": {
                            "type": "string",
                            "format": "string"
                        },
                        "footer": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                }
            ]
        },
        "typeCreate": {
            "type": "object",
            "allOf": [
                {
                    "required": [
                        "code",
                        "translations"
                    ],
                    "properties": {
                        "code": {
                            "type": "string",
                            "format": "string"
                        },
                        "translations": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/typeTranslations"
                            }
                        }
                    }
                }
            ]
        },
        "typeReply": {
            "type": "object",
            "allOf": [
                {
                    "properties": {
                        "id": {
                            "type": "integer",
                            "format": "integer"
                        },
                        "type_key": {
                            "type": "string",
                            "format": "string"
                        },
                        "code": {
                            "type": "string",
                            "format": "string"
                        },
                        "name": {
                            "type": "string",
                            "format": "string"
                        },
                        "created_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "updated_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "deleted_at": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                }
            ]
        },
        "typeDeleteReply": {
            "properties": {
                "string": {
                    "type": "string",
                    "format": "string"
                }
            }
        }
    },
...
}